+++ packages/terra-hookshot/src/HookshotContent.jsx
@@ -190,7 +190,8 @@ class HookshotContent extends React.Component {
     return (
       <div
         {...customProps}
-        className={cx(['content', customProps.className])}
+        // eslint-disable-next-line react/forbid-dom-props
+        className={cx(['content', customProps.className, { 'fixed-position': customProps.hookshotPostionFixed }])}
         ref={(element) => { this.contentNode = element; if (refCallback) { refCallback(element); } }}
       >
         {children}