+++ packages/terra-hookshot/CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased
 
 * Changed
-  * Added conditional class to the HookshotContent component based on the prop `hookshotPostionFixed`
+  * Allow hookshot position to be fixed
 
 ## 5.41.0 - (July 31, 2023)